+++ obj/src/services/FacadeService.d.ts
@@ -3,10 +3,12 @@ import { ConfigParams } from 'pip-services-commons-node';
 import { IReferences } from 'pip-services-commons-node';
 import { IReferenceable } from 'pip-services-commons-node';
 import { CompositeLogger } from 'pip-services-commons-node';
+import { DependencyResolver } from 'pip-services-commons-node';
 import { IFacadeService } from './IFacadeService';
 export declare class FacadeService implements IConfigurable, IReferenceable, IFacadeService {
     protected _rootPath: string;
     protected _partition: any;
+    protected _dependencyResolver: DependencyResolver;
     protected _logger: CompositeLogger;
     constructor();
     configure(config: ConfigParams): void;