+++ obj/src/services/FacadeService.js
@@ -3,10 +3,12 @@ Object.defineProperty(exports, "__esModule", { value: true });
 let _ = require('lodash');
 let express = require('express');
 const pip_services_commons_node_1 = require("pip-services-commons-node");
+const pip_services_commons_node_2 = require("pip-services-commons-node");
 class FacadeService {
     constructor() {
         this._rootPath = '';
         this._partition = express();
+        this._dependencyResolver = new pip_services_commons_node_2.DependencyResolver();
         this._logger = new pip_services_commons_node_1.CompositeLogger();
     }
     configure(config) {
@@ -15,6 +17,7 @@ class FacadeService {
             this._rootPath = '/' + this._rootPath;
     }
     setReferences(references) {
+        this._dependencyResolver.setReferences(references);
         this._logger.setReferences(references);
     }
     getRootPath() {