+++ src/services/PartitionFacadeService.ts
@@ -1,14 +1,12 @@
 import { ConfigParams } from 'pip-services-commons-node';
 import { IReferences } from 'pip-services-commons-node';
 import { Descriptor } from 'pip-services-commons-node';
-import { DependencyResolver } from 'pip-services-commons-node';
 
 import { IFacadeService } from './IFacadeService';
 import { FacadeService } from './FacadeService';
 
 export class PartitionFacadeService extends FacadeService {
     protected _parent: IFacadeService;
-    protected _dependencyResolver: DependencyResolver = new DependencyResolver();
 
     public constructor() {
         super();
@@ -21,13 +19,18 @@ export class PartitionFacadeService extends FacadeService {
     }
 
     public setReferences(references: IReferences): void {
-        this._dependencyResolver.setReferences(references);
+        super.setReferences(references);
         this._parent = this._dependencyResolver.getOneRequired<IFacadeService>('parent');
         this._parent.registerMiddlewareForPath(this._rootPath, this._partition);
+        this.register();
     }
 
     public getRootPath(): string {
         return this._parent.getRootPath() + this._rootPath;
+    }
+
+    protected register(): void {
+        // Override in child classes
     }
 
 }